rename added csv_flavours file (and move out of semi-deprecated fileutils)
